added contacts-controller
