+++ models/contacts.js
@@ -42,14 +42,14 @@ const addContact = async (data) => {
     return newContact;
 }
 
-const updateContact = async (contactId, data) => {
+const updateContact = async (id, data) => {
     const allContacts = await listContacts();
-    const index = allContacts.findIndex(item => item.id === contactId);
+    const index = allContacts.findIndex(item => item.id === id);
     if (index === -1) {
       return null;
     }
-    
-    allContacts[index] = { contactId, ...data };
+  
+    allContacts[index] = { id, ...data };
   
     await updateContacts(allContacts);
     return allContacts[index];
