+++ models/contacts.json
@@ -36,21 +36,33 @@
     "phone": "(186) 568-3720"
   },
   {
-    "contactId": "e6ywwRe4jcqxXfCZOj_1e",
+    "id": "e6ywwRe4jcqxXfCZOj_1e",
     "name": "Nazar",
     "email": "nmosula@gmail.com",
     "phone": "(067) 355-41-41"
   },
   {
-    "contactId": "rsKkOQUi80UsgVPCcLZZW",
+    "id": "rsKkOQUi80UsgVPCcLZZW",
     "name": "Alec2 Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
   },
   {
-    "contactId": "admvhfyCQ5Zu4Uob37on_",
+    "id": "admvhfyCQ5Zu4Uob37on_",
     "name": "Nazar",
     "email": "nmosula@gmail.com",
     "phone": "(067) 355-41-41"
+  },
+  {
+    "id": "cTm3X_g9oXpzsA9ZMxxmb",
+    "name": "Nikhotina Tetiana",
+    "email": "nikhotina@gmail.com",
+    "phone": "(067) 355-41-41"
+  },
+  {
+    "id": "TEmoFuvLaLKPuKmD9mt-v",
+    "name": "Nikhotina Tetiana2",
+    "email": "nikhotina@gmail.com",
+    "phone": "(096) 657-11-52"
   }
 ]
