+++ routes/api/contacts.js
@@ -1,92 +1,18 @@
 const express = require('express')
-const Joi = require('joi')
 
 const router = express.Router()
 
-const contactsService = require("../../models/contacts")
+const contactsController = require('../../controllers/contacts-controllers');
 
-const { HttpError } = require("../../helpers");
 
-const contactAddSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-})
+router.get('/', contactsController.getAllContacts);
 
-router.get('/', async (req, res, next) => {
-  try {
-    const result = await contactsService.listContacts();
-    res.json(result)
-  }
-  catch (error) {
-    next(error);
-  }
-})
+router.get('/:contactId', contactsController.getContactById);
 
-router.get('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contactsService.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404, 'Not found');
-    }
-    res.json(result)
-  }
-  catch (error) {
-    next(error);
-  }
-})
+router.post('/', contactsController.addContact);
 
-router.post('/', async (req, res, next) => {
-  try {
-    const { error } = contactAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(404, error.message);
-    }
-    const result = await contactsService.addContact(req.body);
+router.delete('/:contactId', contactsController.deleteContactById);
 
-    res.status(201).json(result);
-  }
-  catch (error) {
-    next(error);
-  }
-})
-
-router.delete('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contactsService.removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, 'Not found');
-    }
-
-    res.status(200).json({message: "contact deleted"});
-  }
-  catch (error) {
-    next(error);
-  }
-})
-
-router.put('/:contactId', async (req, res, next) => {
-  try {
-    const { error } = contactAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-
-    const { contactId } = req.params;
-
-    const result = await contactsService.updateContact(contactId, req.body);
-
-    if (!result) {
-      throw HttpError(404, 'Not found');
-    }
-
-    res.json(result);
-  }
-  catch (error) {
-    next(error);
-  }
-})
+router.put('/:contactId', contactsController.updateContactById);
 
 module.exports = router
